requirements.txt
requests
beautifulsoup4
cloudscraper
playwright
openai
geopy
flask
python-dotenv
